% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{addIsotopeResult2PeakInfo}
\alias{addIsotopeResult2PeakInfo}
\title{addIsotopeResult2PeakInfo}
\usage{
addIsotopeResult2PeakInfo(
  isotopes.result,
  mz.tol = 25,
  rt.tol = 3,
  int.tol = 500,
  weight.mz = 0.45,
  weight.rt = 0.45,
  weight.int = 0.1,
  peak.info,
  peak.idx,
  anno.idx,
  annotation.idx = c(1:length(peak.info))
)
}
\arguments{
\item{isotopes.result}{The isotope result.}

\item{mz.tol}{The mz tolerance.}

\item{rt.tol}{The RT tolerance.}

\item{int.tol}{The intensity ratio tolerance.}

\item{weight.mz}{mz weight.}

\item{weight.rt}{RT weight.}

\item{weight.int}{Intesnity ratio weight.}

\item{peak.info}{The tags2 data.}

\item{peak.idx}{The index of seeds.}

\item{anno.idx}{The index of annotation}

\item{annotation.idx}{The index of peaks which are annotated.}
}
\value{
Tags2 data.
}
\description{
Add isotope result into tags2 data.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
