% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{addMetaboliteResult2PeakInfo}
\alias{addMetaboliteResult2PeakInfo}
\title{addMetaboliteResult2PeakInfo}
\usage{
addMetaboliteResult2PeakInfo(
  metabolite.result,
  mz.tol = 25,
  rt.tol = 30,
  ccs.tol = 4,
  dp.tol = 0.5,
  matched_frag_tol = 1,
  whether_link_frag = FALSE,
  weight.mz = 0.25,
  weight.rt = 0.25,
  weight.ccs = 0,
  weight.dp = 0.5,
  peak.info,
  peak.idx = 333,
  anno.idx = 1,
  metabolite = metabolite,
  annotation.idx = c(1:length(peak.info))
)
}
\arguments{
\item{metabolite.result}{metabolite.result from metAnnotation}

\item{mz.tol}{m/z tolerrance}

\item{rt.tol}{RT tolerrance}

\item{dp.tol}{Dot product tolerrance}

\item{weight.mz}{m/z weight}

\item{weight.rt}{RT weight}

\item{weight.dp}{Dot product weight}

\item{peak.info}{Peak info data}

\item{peak.idx}{The index of seed peaks.}

\item{anno.idx}{The index of seed annotations.}

\item{metabolite}{KEGG compound database}

\item{annotation.idx}{The index of peaks which are used for annotation}
}
\value{
The tags2 data.
}
\description{
Add annotation information form metAnnotation into peakInfo.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
