% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{addSeedLabel}
\alias{addSeedLabel}
\title{addSeedLabel}
\usage{
addSeedLabel(tags2, round, seed_idx, score_cutoff = 0, matched_frag_tol = 1)
}
\arguments{
\item{tags2}{The tags2 data.}

\item{round}{The annotation round.}

\item{seed_idx}{The index of seeds.}

\item{score_cutoff}{The score cutoff.}
}
\value{
The tags2 data with labeled seeds.
}
\description{
Label seed in tags2 data.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
