% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{annotateCleanMSMS}
\alias{annotateCleanMSMS}
\title{annotateCleanMSMS}
\usage{
annotateCleanMSMS(cleanMSMS, GenForm_an_out, GenForm_loss_out, lowestPpm=TRUE, round=TRUE, removeAnnoSpace=FALSE)
}
\arguments{
\item{cleanMSMS}{File name for Cleaned MSMS text file output from \code{\link{RunGenForm}}}

\item{GenForm_an_out}{File name for annotated MSMS text file output from \code{\link{RunGenForm}}}

\item{GenForm_loss_out}{File name for annotated "loss" MSMS text file output from \code{\link{RunGenForm}}}

\item{lowestPpm}{If \code{TRUE}, choses the lowest ppm subformula for assignment of formula
to fragments. If \code{FALSE}, takes the first formula.}

\item{round}{If \code{TRUE}, rounds the output to 4 (mz) and 1 (Int) decimal places for aesthetics.
If \code{FALSE}, no rounding is done.}

\item{removeAnnoSpace}{If \code{TRUE}, removes the space in the \code{"no loss"} annotation.}
}
\value{
Returns annotated peaks for further use
}
\description{
This takes the output from \code{\link{RunGenForm}} and creates a peaklist with the annotations for further use.
}
\examples{

}
\author{
Emma Schymanski (\code{\link{emma.schymanski@uni.lu}})
}
