% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{annotateInitialSeed}
\alias{annotateInitialSeed}
\title{annotateInitialSeed}
\usage{
annotateInitialSeed(
  ms1_file = "data.csv",
  ms2_file = NULL,
  metdna_version = c("version2", "version1"),
  ms2_type = c("mgf", "mzXML", "msp", "cef"),
  path = ".",
  polarity = c("positive", "negative"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "WatersQTOF", "WatersTWIMMS", "AgilentDTIMMS", "BrukerTIMS"),
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  column = c("hilic", "rp"),
  ce = c("30", "10", "20", "35,15", "40", "50", "NCE10", "NCE20", "NCE30", "NCE40",
    "NCE50", "SCE20_30_40\%", "SNCE20_30_40\%"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min"),
  excluded_adduct = NULL,
  is_rt_calibration = FALSE,
  mz_tol = 25,
  mz_ppm_thr = 400,
  pf_rt_range = 0,
  tolerance_rt_range = 30,
  pf_ccs_range = 0,
  tolerance_ccs_range = 2,
  is_filter = TRUE,
  is_rt_score = TRUE,
  is_ccs_score = FALSE,
  is_ms2_score = TRUE,
  is_include_precursor = TRUE,
  int_ms2_min_abs = 50,
  int_ms2_min_relative = 0.01,
  mz_tol_combine_ms1_ms2 = 25,
  rt_tol_combine_ms1_ms2 = 10,
  ccs_tol_combine_ms1_ms2 = NULL,
  mz_tol_ms2 = 35,
  dp_cutoff = 0.8,
  matched_frag_cutoff = 1,
  direction = c("reverse", "forward"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  test_adduct_version = c("version2", "version1"),
  test_evaluation = c("No", "200STD", "46STD"),
  test_force_filtering_rt = NULL,
  is_plot_ms2 = TRUE
)
}
\arguments{
\item{ms1_file}{The name of ms1 peak table. Column 1 is "name", Column 2 is "mz" and column is "rt".}

\item{ms2_file}{Default: NULL}

\item{metdna_version}{'version1', 'version2'. Default: "version1"}

\item{ms2_type}{"mgf", "mzXML", "msp", "cef"}

\item{path}{Default: '.'}

\item{instrument}{The instrument you used to acquire data. "SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", 'WatersQTOF','WatersTWIMMS'. Default: "SciexTripleTOF"}

\item{lib}{Default: 'zhuMetLib'}

\item{column}{"hilic", "rp"}

\item{ce}{"10", "20", "30", "35,15", "40", "50". Default: "30"}

\item{method_lc}{'Amide12min', 'Amide23min', 'Other', 'MetlinRP', 'RP12min'. Default: 'Amide12min'}

\item{excluded_adduct}{adduct list for exclusion. Default: NULL}

\item{is_rt_calibration}{Default: FALSE}

\item{mz_tol}{ms1 match. Default: 25 ppm}

\item{tolerance_rt_range}{maxmium rt tolerance. Default: 30 s}

\item{tolerance_ccs_range}{maxmium rt tolerance. Default: 2 %}

\item{is_filter}{Whether filter candidates whose rt/ccs exceed the maximum tolerance. Default: FALSE}

\item{is_rt_score}{logical vector. Default: TRUE}

\item{is_ccs_score}{logical vector. Default: FALSE}

\item{is_ms2_score}{logical vector. Default: TRUE}

\item{is_include_precursor}{Default: TRUE}

\item{int_ms2_min_abs}{Default: 30}

\item{int_ms2_min_relative}{Default: 0.01}

\item{mz_tol_combine_ms1_ms2}{Default: 25 ppm}

\item{rt_tol_combine_ms1_ms2}{Default: 10 s}

\item{ccs_tol_combine_ms1_ms2}{Only appliable for IM-MS data. Default: NULL}

\item{mz_tol_ms2}{Default: 35 ppm}

\item{dp_cutoff}{The cutoff of dot product. Default is 0.8.}

\item{matched_frag_cutoff}{The cutoff of matched fragment number. Default: 1}

\item{direction}{Reserve which direct dot product. 'reverse', 'forward'. Default: 'reverse'}

\item{scoring_approach}{spectral match approach, including 'dp', 'bonanza', 'hybrid', 'gnps'. Default: 'dp'.}

\item{is_plot_ms2}{Output MS2 match plot or not. Default: TRUE}

\item{pf_rt_tol}{penalty-free rt range. Default: 0 s}

\item{pf_ccs_tol}{penalty-free rt range. Default: 0 %}
}
\value{
Return the annotation result.
}
\description{
Annotate peak using MS/MS spectra in in-house database.
}
\author{
Zhiwei Zhou, Xiaotao Shen, Yandong Yin
}
