% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{annotateIsotopeMRN}
\alias{annotateIsotopeMRN}
\title{annotateIsotopeMRN}
\usage{
annotateIsotopeMRN(
  id = NA,
  formula = "C21H27N7O14P2",
  adduct = "M+H",
  charge = 1,
  mz = 664.1139,
  rt = 812.143,
  int = 205109.4,
  peak_mz_all,
  peak_rt_all,
  peak_int_all,
  peak_cor_all,
  rt_tol = 5,
  mz_tol = 25,
  mz_ppm_thr = 400,
  cor_tol = 0,
  int_tol = 500,
  max_isotope = 4
)
}
\arguments{
\item{id}{The KEGG ID of metabolite.}

\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct of metabolite.}

\item{charge}{The charge of metabolite.}

\item{mz}{The mz of metabolite.}

\item{rt}{The rt of metabolite.}

\item{int}{The mean/median intensity of metabolite accross the samples.}

\item{peak.mz}{The mz of peaks.}

\item{peak.rt}{The rt of peaks.}

\item{peak.int}{The mean/median intensity of peaks accross the samples.}

\item{cor}{The correlations between metabolite and peaks.}

\item{rt.tol}{The rt tolerance (second).}

\item{mz.tol}{The mz tolerance.}

\item{cor.tol}{The cor tolerance.}

\item{int.tol}{The intensity ratio tolerance.}

\item{max.isotope}{The max number of isotopes.}
}
\value{
Isotope annotation information of peaks.
}
\description{
Find the isotopes of a known metabolite according to rt, mz and intensity. Old function in MetDNA1.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
