% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{annotateMetaboliteMRN}
\alias{annotateMetaboliteMRN}
\title{annotateMetaboliteMRN}
\usage{
annotateMetaboliteMRN(
  metabolite_name = "M175T962",
  metabolite_id = "C00062",
  formula = "C6H14N4O2",
  adduct = "M+H",
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  mz = 175.118,
  rt = 961.6225,
  peak_mz_all,
  peak_rt_all,
  peak_ccs_all,
  ms2,
  mz_tol = 25,
  mz_ppm_thr = 400,
  rt_tol = 30,
  ccs_tol = 4,
  dp_tol = 0.5,
  matched_frag_tol = 1,
  whether_link_frag = FALSE,
  step = 1,
  mz_tol_ms2 = 25,
  whether_use_predRT = TRUE,
  metabolite,
  metabolite_ccs,
  metabolic_network,
  adduct_table,
  ...
)
}
\arguments{
\item{formula}{The formula of metabolite.}

\item{adduct}{The adduct of metabolite.}

\item{mz}{The mz of metabolite.}

\item{rt}{The RT of metabolite.}

\item{ms2}{The ms2 data of peak table.}

\item{step}{The reaction step.}

\item{metabolite}{The kegg compound database.}

\item{metabolite.name}{The metabolite name.}

\item{metabolite.id}{The metabolite ID.}

\item{polarity}{The polarity.}

\item{peak.mz}{The mz of all peaks.}

\item{peak.rt}{The RT of all peaks.}

\item{cor}{The correlation of metabolite between all peaks.}

\item{mz.tol}{The mz tolerance.}

\item{rt.tol}{The RT tolerance for metabolite annotation. (\%)}

\item{cor.tol}{The cor tolerance.}

\item{dp.tol}{The tolerance of dot product.}

\item{metabolic.network}{kegg.rpair2}

\item{adduct.table}{Adduct table.}
}
\value{
The metabolite annotation information.
}
\description{
Annotate peak table from one metabolite.
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
