% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{annotateMRN}
\alias{annotateMRN}
\title{annotateMRN}
\usage{
annotateMRN(
  annotation_result = "ms2_match_annotation_result.csv",
  ms2_data = "ms2",
  prefer_adduct = c("all", "M+H", "M+Na", "M-H"),
  use_default_md = TRUE,
  use_pretrained_model = FALSE,
  metdna_version = c("version2", "version1"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  direction = c("reverse", "forward"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  column = c("hilic", "rp"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min"),
  excluded_adduct = NULL,
  polarity = c("positive", "negative"),
  extension_step = c("0", "1", "2", "3", "4", "5", "6", "7", "8"),
  threads = 3,
  path = ".",
  max_isotope = 4,
  mz_tol = 25,
  mz_ppm_thr = 400,
  rt_tol1 = 3,
  rt_tol2 = 30,
  ccs_tol = 4,
  cor_tol = 0,
  int_tol = 500,
  dp_tol = 0.5,
  matched_frag_tol = 1,
  whether_link_frag = FALSE,
  max_step = 3,
  score_cutoff = 0,
  remain = FALSE,
  remain_per = 0.5,
  seed_neighbor_match_plot = TRUE,
  candidate_num = 3,
  test_adduct_version = c("version2", "version1"),
  test_evaluation = c("No", "200STD", "46STD"),
  whether_use_predRT = TRUE,
  ...
)
}
\arguments{
\item{annotation_result}{"ms2_match_annotation_result.csv"}

\item{ms2_data}{'ms2'}

\item{prefer_adduct}{prefer adduct for RT prediction, including "all", "M+H", "M+Na", "M-H". Default: 'M+H'}

\item{use_default_md}{Whether use default MD.}

\item{metdna_version}{MetDNA version, 'version1', 'version2'. Default: 'version2'}

\item{direction}{'forward', 'reverse'. Default: 'reverse'}

\item{column}{'hilic', 'rp'}

\item{method_lc}{'Amide12min', 'Amide23min', 'Other', 'MetlinRP', 'RP12min'}

\item{excluded_adduct}{NULL}

\item{polarity}{'positive', 'negative'. Default: 'positive'}

\item{extension_step}{The used extended MRN with n step reaction. Steps: '0', '1', '2', '3', '4', '5', '6', '7', '8'}

\item{threads}{Default: 4}

\item{path}{'.'}

\item{max_isotope}{Maximum considerable isotope. Default: 4}

\item{mz_tol}{Default: 25 ppm}

\item{rt_tol1}{RT tolerance of adductAnnotation and isotopeAnnotation. Default: 3s}

\item{rt_tol2}{RT tolerance of metAnnotation. Default: 30%}

\item{cor_tol}{Default: 0}

\item{int_tol}{isotope intensity tolerance. Defautl: 500}

\item{dp_tol}{0.5}

\item{max_step}{the maximum step of reaction pair during recursive annotation}

\item{score_cutoff}{Total score cutoff. Default: 0}

\item{remain}{FALSE}

\item{remain_per}{0.5}

\item{seed_neighbor_match_plot}{TRUE}

\item{candidate_num}{reserve top N candidate number. Default: 3}

\item{Use}{seed from reserved match or forward match. 'forward', 'reverse'. Default: 'reverse'}
}
\value{

}
\description{
metabolic reaction network based recursive annotation
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
