% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{annotateNeighbor}
\alias{annotateNeighbor}
\title{annotateNeighbor}
\usage{
annotateNeighbor(
  data = "ms2_match_annotation_result.csv",
  ms2,
  adduct_table,
  max_isotope = 4,
  polarity = c("positive", "negative"),
  metdna_version = c("version2", "version1"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min"),
  direction = c("forward", "reverse"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  mz_tol = 25,
  mz_ppm_thr = 400,
  rt_tol1 = 3,
  rt_tol2 = 30,
  cor_tol = 0,
  int_tol = 500,
  ccs_tol = 4,
  dp_tol = 0.5,
  matched_frag_tol = 1,
  whether_link_frag = FALSE,
  max_step = 3,
  remain = FALSE,
  remain_per = 0.5,
  metabolite,
  metabolite_ccs,
  metabolic_network,
  inHouse_compound,
  threads = 3,
  score_cutoff = 0,
  path = ".",
  rt_filter = TRUE,
  seed_neighbor_match_plot = FALSE,
  candidate_num = 3,
  test_evaluation = c("No", "200STD", "46STD"),
  whether_use_predRT = TRUE,
  ...
)
}
\arguments{
\item{data}{The data (csv) from ms2Annotation.}

\item{ms2}{The ms2 data of peak table.}

\item{polarity}{The polarity.}

\item{remain}{Remain some seeds as validation or not.}

\item{metabolite}{The kegg compound database.}

\item{threads}{The number of threads.}

\item{path}{The directory.}

\item{adduct.table}{Adduct table.}

\item{max.isotope}{The number of isotope peaks.}

\item{mz.tol}{The mz tolerance.}

\item{rt.tol1}{The RT tolerance for isotope and adduct annotation. (second)}

\item{rt.tol2}{The RT tolerance for metabolite annotation. (\%)}

\item{cor.tol}{The cor tolerance.}

\item{int.tol}{The intensity ratio tolerance.}

\item{dp.tol}{The tolerance of dot product.}

\item{max.step}{The max number of reaction step.}

\item{remain.per}{The percentage of remained seeds.}

\item{metabolic.network}{kegg.rpair2.}

\item{inHouse.compound}{The inhouse compound database with predicted RT.}

\item{score.cutoff}{Score cutoff.}

\item{output.path}{The directory of outputing results.}
}
\value{
The tags data with annotation result.
}
\description{
Annotate peak table from seeds.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
