% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{annotatePeakGroup}
\alias{annotatePeakGroup}
\title{annotatePeakGroup}
\usage{
annotatePeakGroup(
  peak_group,
  ms2_data,
  polarity = c("positive", "negative"),
  tol_mz = tol_mz,
  is_ms2_check = TRUE,
  ms2_score_cutoff = -1,
  lib_adduct_nl = lib_adduct_nl,
  cutoff_ssc = 0.3,
  cutoff_ssc_int = 3000,
  is_rule_limitation = TRUE,
  cutoff_topN = 5,
  ...
)
}
\arguments{
\item{peak_group}{a object of peak_group}

\item{tol_mz}{Default: 10 ppm}

\item{is_ms2_check}{whether compare ms2 similarity of neutral loss to base peak ms2. Default: TRUE}

\item{ms2_score_cutoff}{Default: -1; # -1 represent not filter}

\item{cutoff_ssc}{Whether apply ssc filter in annotation. Default: 0.3}

\item{cutoff_ssc_int}{If the peak intensity less than this value, cutoff_ssc_int is invalid.}

\item{is_rule_limitation}{Whether apply rule limitation in adductAnnotation and nlAnnotation (e.g. '[2M+H]+' needs '[M+H]+'). Default: TRUE}

\item{cutoff_topN}{Top N fragments in base peak ms2 used for ISF annotation. Default: 5}
}
\description{
annotate peak group with isotopes, adducts, neutral loss and in-source fragments
}
\author{
Zhiwei Zhou
}
