% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{assignConfidence}
\alias{assignConfidence}
\title{assignConfidence}
\usage{
assignConfidence(x, polarity = c("positive", "negative"))
}
\arguments{
\item{x}{The metabolite ID group.}
}
\value{
The confidence of metabolite ID group.
}
\description{
Assign confidence to metabolite ID group.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
