% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{authenticateAnnotationCredential}
\alias{authenticateAnnotationCredential}
\title{authenticateAnnotationCredential}
\usage{
authenticateAnnotationCredential(
  peak_table_file,
  annotation_initial_file,
  ms2_data_file = "ms2_data.RData",
  path_dir = ".",
  thread = 4,
  tol_rt = 3,
  polarity = c("positive", "negative"),
  tol_mz = 25,
  isotope_int_ratio_check = TRUE,
  isotope_int_ratio_cutoff = 500,
  is_ms2_check = TRUE,
  ms2_score_cutoff = -1,
  cutoff_ssc = 0.3,
  cutoff_ssc_int = 3000,
  is_rule_limitation = TRUE,
  cutoff_topN = 5,
  is_plot_pseudo_MS1 = TRUE,
  type_order = c("level1", "level2", "level3"),
  is_pred_formula_all = FALSE,
  dir_GenForm = "/home/zhouzw/software/genform-code-r8-trunk/src/GenForm/bin/Debug",
  ppm = 10,
  acc = 15,
  elements = "CHNOPS",
  num_formula_candidate = 3,
  platform = c("windows", "linux")
)
}
\arguments{
\item{peak_table_file}{'peak_table.csv'}

\item{annotation_initial_file}{'annotation_initial.csv'}

\item{ms2_data_file}{'ms2_data.RData'}

\item{path_dir}{path of working dictory. Default: '.'}

\item{thread}{number of thread. Default: 4}

\item{tol_rt}{rt tolerance for peak group generation, Default: 3s}

\item{polarity}{ionzation polarity, 'positive', 'negative'; Default: 'positive'}

\item{tol_mz}{mz tolerance. Default: 10 ppm}

\item{isotope_int_ratio_check}{whether check isotope intensity; Default: TRUE}

\item{isotope_int_ratio_cutoff}{isotope intensity ratio cutoff; Default: 500%}

\item{is_ms2_check}{whether compare ms2 of NL with base peak; Default: TRUE}

\item{ms2_score_cutoff}{Default: -1; # -1 represent not filter}

\item{cutoff_ssc}{sample-sample correlation for peak group generation, Default: 0.3}

\item{is_plot_pseudo_MS1}{whether output the pseudo MS1 spectrum.}

\item{is_pred_formula_all}{whether predict formula for all peak group (TRUE) or concised peak groups (FALSE). Default: FALSE}

\item{dir_GenForm}{directory path of GenForm}

\item{ppm}{m/z tolerance of MS1; Default: 10 ppm}

\item{acc}{m/z tolerance of MS2; Default: 15 ppm}

\item{elements}{CHNOPS}

\item{num_formula_candidate}{3}
}
\description{
authenticate annotations with semi-targeted feature clustering and formula prediction.
}
\examples{
peak_table_file <- system.file("extdata", "peak_table.csv", package="MetDNA2")
annotation_initial_file <- system.file("extdata", "annotation_initial.csv", package="MetDNA2")
ms2_data <- system.file("tempdata", "ms2_data.RData", package="MetDNA2")
authenticateAnnotationCredential(peak_table_file = "peak_table.csv",
                                 annotation_initial_file = "annotation_initial.csv",
                                 ms2_data = "ms2_data.RData",
                                 path_dir = 'I:/00_projects/03_MetDNA2/00_data/20200811_annotationCredential_workflow/demo_data2',
                                 polarity = 'negative')
}
\author{
Zhiwei Zhou
}
