% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{calculateExportTableScore}
\alias{calculateExportTableScore}
\title{calculateExportTableScore}
\usage{
calculateExportTableScore(
  table_identification,
  mz_tol = 25,
  rt_tol = 30,
  ccs_tol = 3,
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  ...
)
}
\description{
calculateExportTableScore
}
\author{
Zhiwei Zhou
}
