% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{calculateRedundancy}
\alias{calculateRedundancy}
\title{calculateRedundancy}
\usage{
calculateRedundancy(result)
}
\arguments{
\item{result}{The result from metA.}
}
\value{
peak and metabolite redundancy.
}
\description{
Calculate redundancy of result from metA.
}
\author{
Xiaotao Shen
}
