% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImporting.R
\name{checkPara}
\alias{checkPara}
\title{checkPara}
\usage{
checkPara(
  path,
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  polarity = c("positive", "negative"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  column = c("hilic", "rp"),
  ce = c("30", "10", "20", "35,15", "40", "50", "NCE10", "NCE20", "NCE30", "NCE40",
    "NCE50", "SCE20_30_40\%", "SNCE20_30_40\%"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min"),
  is_rt_calibration = TRUE
)
}
\description{
check parameters
}
\author{
Zhiwei Zhou
}
