% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImporting.R
\name{checkQuality}
\alias{checkQuality}
\title{checkQuality}
\usage{
checkQuality(
  ms1_file = "data.csv",
  sample_info_file = "sample.info.csv",
  ms2_type = c("mgf", "mzXML", "msp", "cef"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  path = "."
)
}
\arguments{
\item{ms1_file}{The name of ms1 ms1_file.}

\item{sample_info_file}{The name of sample information.}

\item{ms2_type}{The type of MS2 ms1_file.}

\item{path}{The work directory.}
}
\value{
Check result.
}
\description{
Check ms1_file.
}
\author{
Xiaotao Shen, Zhiwei Zhou
}
