% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImporting.R
\name{checkQualitySampleGroup}
\alias{checkQualitySampleGroup}
\title{checkQualitySampleGroup}
\usage{
checkQualitySampleGroup(
  ms1_file = "data.csv",
  sample_info_file = "sample.info.csv",
  group,
  path = "."
)
}
\arguments{
\item{ms1_file}{"data.csv"}

\item{sample_info_file}{"sample.info.csv"}

\item{path}{'.'}
}
\description{
checkQualitySampleGroup
}
\author{
Zhiwei Zhou
}
