% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{combineMs1Ms2}
\alias{combineMs1Ms2}
\title{combineMS1MS2}
\usage{
combineMs1Ms2(
  ms1_data,
  ms2_data,
  mz_tol_combine_ms1_ms2 = 25,
  rt_tol_combine_ms1_ms2 = 10,
  ccs_tol_combine_ms1_ms2 = NULL,
  path = ".",
  ms2_type = c("mgf", "mzXML", "msp", "cef")
)
}
\arguments{
\item{path}{Directory.}

\item{ms1.file}{The name of ms1 peak table. Column 1 is "name", Column 2 is "mz" and column 3 is "rt".}

\item{ms2.file}{The vector of names of ms2 files. MS2 file must be mzXML.}

\item{mz.tol}{mz tol for ms1 and ms2 data matching.}

\item{rt.tol}{RT tol for ms1 and ms2 data matching.}

\item{ms2.type}{The type of MS2 file, default is mzXML.}
}
\value{
Return ms1 and ms2 data.
}
\description{
Combine MS1 and MS2 data.
}
\author{
Xiaotao Shen, Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
