% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{concisePeak2PeakGroup}
\alias{concisePeak2PeakGroup}
\title{concisePeak2PeakGroup}
\usage{
concisePeak2PeakGroup(
  list_peak_group_annotation,
  path_dir,
  type_order = c("level1", "level2", "level3"),
  ...
)
}
\description{
concise conflict peak groups (one peak - multiple peak groups(i.e., adduct), M167T834_[M-H]- & M167T834_[M-H2O-H]-)
}
\examples{
load(system.file("tempdata", "list_peak_group_annotation_200805.RData", package="MetDNA2"))
test <- concisePeak2PeakGroup(list_peak_group_annotation = list_peak_group_annotation)
}
\author{
Zhiwei Zhou
}
