% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{concisePeakGroup2Annotation}
\alias{concisePeakGroup2Annotation}
\title{concisePeakGroup2Annotation}
\usage{
concisePeakGroup2Annotation(list_peak_group_annotation)
}
\description{
convert peak groups to annotation table
}
\examples{
load(system.file("tempdata", "list_peak_group_annotation_concised_pg2pg_200805.RData", package="MetDNA2"))
test <- concisePeakGroup2Annotation(list_peak_group_annotation_concised)
}
\author{
Zhiwei Zhou
}
