% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{concisePeakGroup2PeakGroup}
\alias{concisePeakGroup2PeakGroup}
\title{concisePeakGroup2PeakGroup}
\usage{
concisePeakGroup2PeakGroup(
  list_peak_group_annotation,
  type_order = c("level1", "level2", "level3"),
  path_dir,
  ...
)
}
\description{
concise overlap peak groups (Peak group - peak group, e.g. M175T752_[M-H]-, ISF of M175T752 (M132T752_[M+NH4-2H]-, M132T752_[M-H]-), Isotope of M175T752 (M177T752_[M-H]-))
}
\examples{
load(system.file("tempdata", "list_peak_group_annotation_concised_p2pg_200805.RData", package="MetDNA2"))
test <- concisePeak2PeakGroup(list_peak_group_annotation = list_peak_group_annotation_concised)
}
\author{
Zhiwei Zhou
}
