% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{convertAdductList2AdductTable}
\alias{convertAdductList2AdductTable}
\title{convertAdductList2AdductTable}
\usage{
convertAdductList2AdductTable(
  adduct_list = NULL,
  polarity = c("positive", "negative"),
  column = c("hilic", "rp")
)
}
\arguments{
\item{adduct_list}{Default: NULL}

\item{polarity}{'positive', 'negative'}

\item{column}{'hilic', 'rp'}
}
\description{
convert adduct_list to adduct_table (MetDNA format)
}
\author{
Zhiwei Zhou
}
