% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{convertAnnotationTable2InitialId}
\alias{convertAnnotationTable2InitialId}
\title{convertAnnotationTable2InitialId}
\usage{
convertAnnotationTable2InitialId(
  peak_table_file = "idresults.csv",
  ms2_file = "spectra.msp",
  sample_info_file = "sample.info.csv",
  path_dir = ".",
  lib = c("zhulib", "zhumetlib", "kegg"),
  match_method = c("forward", "reverse"),
  polarity = c("positive", "negative"),
  tool = c("MetAnalyzer1", "MetAnalyzer2", "MetDNA", "MetDNA2"),
  ...
)
}
\arguments{
\item{peak_table_file}{Default: 'idresults.csv'}

\item{ms2_file}{'spectra.msp'}

\item{sample_info_file}{'sample.info.csv'}

\item{path_dir}{Default: '.'}

\item{lib}{library of IDs. 'zhulib', 'zhumetlib', 'kegg'. Default: 'zhumetlib'}

\item{match_method}{'forward', 'reverse'; Default: 'reverse'}

\item{polarity}{'positive', 'negative'}

\item{tool}{'MetAnalyzer1', 'MetAnalyzer2', 'MetDNA'; Default:'MetAnalyzer1'}
}
\description{
convert Annotation Table (MetAnalyzer) to initial ID table
}
\author{
Zhiwei Zhou
}
