% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Class.R
\name{convertMs1Data2SpecAnnotationClass}
\alias{convertMs1Data2SpecAnnotationClass}
\title{convertMs1Data2SpecAnnotationClass}
\usage{
convertMs1Data2SpecAnnotationClass(ms1_info)
}
\description{
Convert Ms1DataInfor to AnnotationResult class
}
\author{
Zhiwei Zhou
}
