% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{convertMs2RData}
\alias{convertMs2RData}
\title{convertMs2RData}
\usage{
convertMs2RData(ms2_file, path_dir = ".")
}
\arguments{
\item{path_dir}{'.'}
}
\description{
convertMs2RData
}
\examples{
convertMs2RData(ms2_file = 'spectra.msp',
                path_dir = 'I:/00_projects/03_MetDNA2/00_data/20200811_annotationCredential_workflow/demo_data')
}
\author{
Zhiwei Zhou
}
