% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{convertResultMatrix2Tags}
\alias{convertResultMatrix2Tags}
\title{convertResultMatrix2Tags}
\usage{
convertResultMatrix2Tags(result, tags2 = tags2)
}
\arguments{
\item{result}{The result.}

\item{tags2}{The tags2 data.}
}
\value{
The tags2 data.
}
\description{
Filter tags2 data according to result.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
