% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{convertTags2ResultMatrix}
\alias{convertTags2ResultMatrix}
\title{convertTags2ResultMatrix}
\usage{
convertTags2ResultMatrix(tags2, score.cutoff = 0, candidate.num = 3000)
}
\arguments{
\item{tags2}{The tags2 data.}

\item{score.cutoff}{The score cutoff.}

\item{candidate.num}{The number of candidates.}
}
\value{
The matrix of tags2 data.
}
\description{
Transform tags2 data to matrix like ms2Annotation.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
