% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{featureReName}
\alias{featureReName}
\title{featureReName}
\usage{
featureReName(name)
}
\arguments{
\item{name}{a vector. character.}
}
\description{
define feature name of replicates.
}
\author{
Zhiwei Zhou
}
