% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateAnnotationSummary4Report}
\alias{generateAnnotationSummary4Report}
\title{generateAnnotationSummary4Report}
\usage{
generateAnnotationSummary4Report(result_annotation)
}
\description{
generate annotation summary for report
}
\author{
Zhiwei Zhou
}
