% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateBioInterpSummary4Report}
\alias{generateBioInterpSummary4Report}
\title{generateBioInterpSummary4Report}
\usage{
generateBioInterpSummary4Report(
  result_stat_feature,
  result_pathway,
  p_cutoff = 0.05,
  ...
)
}
\description{
generateBioInterpSummary4Report
}
\author{
Zhiwei Zhou
}
