% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateCredentialSummary4Report}
\alias{generateCredentialSummary4Report}
\title{generateCredentialSummary4Report}
\usage{
generateCredentialSummary4Report(
  result_credential_initial,
  result_credential,
  result_credential_pg,
  result_credential_filter
)
}
\description{
generateCredentialSummary4Report
}
\author{
Zhiwei Zhou
}
