% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{generateExporTableIden}
\alias{generateExporTableIden}
\title{generateExporTableIden}
\usage{
generateExporTableIden(
  annot_all,
  peak_group_id_table,
  direction = c("reverse", "forward"),
  tolerance_rt_range = 30,
  lib = "zhuMetLib",
  test_evaluation = c("No", "200STD", "46STD"),
  ...
)
}
\description{
generateExporTableIden
}
\author{
Zhiwei Zhou
}
