% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{generateExportTable}
\alias{generateExportTable}
\title{generateExportTable}
\usage{
generateExportTable(
  annot_all,
  path = ".",
  direction = c("reverse", "forward"),
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  tolerance_rt_range = 30,
  is_cred_formula_filter = FALSE,
  candidate_num = 10,
  test_evaluation = c("No", "200STD", "46STD"),
  ...
)
}
\arguments{
\item{path}{'.'}

\item{direction}{'reverse', 'forward'}

\item{lib}{c('zhuMetLib', 'zhuMetLib_orbitrap', 'fiehnHilicLib')}

\item{instrument}{c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", 'ThermoExploris', "AgilentDTIMMS", "BrukerTIMS", 'WatersQTOF','WatersTWIMMS')}

\item{tolerance_rt_range}{30}

\item{is_cred_formula_filter}{Default: FALSE}

\item{candidate_num}{5}
}
\description{
generateExportTable
}
\author{
Zhiwei Zhou
}
