% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{generateExportTablePeakGroup}
\alias{generateExportTablePeakGroup}
\title{generateExportTablePeakGroup}
\usage{
generateExportTablePeakGroup(annot_all, peak_group_id_table)
}
\description{
generateExportTablePeakGroup
}
\author{
Zhiwei Zhou
}
