% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{generateFiles4FormulaPrediction}
\alias{generateFiles4FormulaPrediction}
\title{generateFiles4FormulaPrediction}
\usage{
generateFiles4FormulaPrediction(
  peak_group,
  ms2_spec,
  polarity = c("positive", "negative"),
  path_dir = "."
)
}
\arguments{
\item{polarity}{c('positive', 'negative')}

\item{path_dir}{'.'}
}
\value{
a object of PeakGroupFormula
}
\description{
Generate files (ms1_file & ms2_file) for formula prediction
}
\examples{
load('./inst/tempdata/list_peak_group_annotation_concised_pg2pg_200805.RData')
load('./inst/tempdata/raw_msms_200805.RData')
peak_group <- list_peak_group_annotation_concised$`M482T929_[M-H]-`
polarity <- 'negative'
ms2_spec <- raw_msms$M482T929$spec
path_dir <- 'I:/00_projects/03_MetDNA2/00_data/20200810_conbine_Genform_to_feature_credential/demo_formula'
test <- generateFiles4FormulaPrediction(peak_group = peak_group,
                                        ms2_spec = ms2_spec,
                                        polarity = 'negative',
                                        path_dir = path_dir)
}
\author{
Zhiwei Zhou
}
