% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{generateFormulaResult}
\alias{generateFormulaResult}
\title{generateFormulaResult}
\usage{
generateFormulaResult(list_peak_group_formula, num_formula_candidate = 5)
}
\arguments{
\item{num_formula_candidate}{top N formula reserved. Default: 5}
}
\description{
generateFormulaResult
}
\examples{
load('./inst/tempdata/list_peak_group_formula.RData')
test <- generateFormulaResult(list_peak_group_formula, num_formula_candidate = 3)
}
\author{
Zhiwei Zhou
}
