% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMerge.R
\name{generateMetDNA2AnnoClass}
\alias{generateMetDNA2AnnoClass}
\title{generateMetDNA2AnnoClass}
\usage{
generateMetDNA2AnnoClass(result_initial_annotation)
}
\description{
Convert Ms1DataInfor to AnnotationResult class
}
\author{
Zhiwei Zhou
}
