% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Class.R
\name{generateMetDNA2Annotation}
\alias{generateMetDNA2Annotation}
\title{generateMetDNA2Annotation}
\usage{
generateMetDNA2Annotation(result_initial_annotation)
}
\description{
Convert Ms1DataInfor to AnnotationResult class
}
\author{
Zhiwei Zhou
}
