% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{generateMetDNA2AnnotationResult}
\alias{generateMetDNA2AnnotationResult}
\title{generateMetDNA2AnnotationResult}
\usage{
generateMetDNA2AnnotationResult(
  path = ".",
  thread = 4,
  is_cred_pg_filter = TRUE,
  is_cred_formula_filter = TRUE,
  is_pred_formula_all = FALSE,
  direction = c("reverse", "forward"),
  tolerance_rt_range = 30,
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  remove_conflict_seed_from_final_table = FALSE,
  test_rm_extra_anno_from_ini_seed = FALSE,
  ...
)
}
\arguments{
\item{path}{'.'}

\item{thread}{Default: 4}
}
\description{
generateMetDNA2AnnotationResult
}
\author{
Zhiwei Zhou
}
