% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateMetDNA2Report}
\alias{generateMetDNA2Report}
\title{generateMetDNA2Report}
\usage{
generateMetDNA2Report(
  sample_info = "sample.info.csv",
  path = ".",
  polarity = c("positive", "negative", "both"),
  export_type = c("html", "pdf", "all"),
  extension_step = c("0", "1", "2", "3", "4", "5", "6", "7", "8"),
  is_rt_calibration = TRUE,
  is_bio_interpret = TRUE
)
}
\description{
generateMetDNA2Report
}
\author{
Zhiwei Zhou
}
