% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Tools.R
\name{generateMSP}
\alias{generateMSP}
\title{generateMSP}
\usage{
generateMSP(
  file_name = "./test.msp",
  cmp_name = NULL,
  precusormz = NULL,
  adduct = NULL,
  instrument_type = NULL,
  instrument = NULL,
  smiles = NULL,
  inchikey = NULL,
  inchikey1 = NULL,
  formula = NULL,
  polarity = c("positive", "negative"),
  ce = NULL,
  rt = NULL,
  ccs = NULL,
  zhulib_id = NULL,
  kegg_id = NULL,
  hmdb_id = NULL,
  pubchem_id = NULL,
  links = "",
  comment = "",
  spec = NULL
)
}
\arguments{
\item{file_name}{Required. The file name of MSP. The suffix of file name must be ".msp".}

\item{cmp_name}{Required}

\item{precusormz}{Required}

\item{adduct}{Default: NULL}

\item{instrument_type}{Default: NULL}

\item{instrument}{Default: NULL}

\item{smiles}{Default: NULL}

\item{inchikey}{Default: NULL}

\item{inchikey1}{Default: NULL}

\item{formula}{Default: NULL}

\item{polarity}{'Positive' or 'Negative'}

\item{ce}{Default: NULL}

\item{rt}{Default: NULL}

\item{ccs}{Default: NULL}

\item{zhulib_id}{Default: NULL}

\item{kegg_id}{Default: NULL}

\item{hmdb_id}{Default: NULL}

\item{pubchem_id}{Default: NULL}

\item{links}{Default: ''}

\item{comment}{Default: ''}

\item{spec}{Required. It should be in a dataframe, 1st: "mz", 2nd: "intensity"}
}
\description{
Convert MS/MS spectra to MSP files
}
\author{
Zhiwei Zhou
}
