% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateParaList4Report}
\alias{generateParaList4Report}
\title{generateParaList4Report}
\usage{
generateParaList4Report(
  para_file = "para_list.txt",
  path = ".",
  showed_para = c("package_version", "ms1_file", "ms2_file", "sample_info_file", "lib",
    "polarity", "instrument", "column", "ce", "method_lc", "is_check_data",
    "is_anno_initial_seed", "is_anno_mrn", "is_credential", "mz_tol", "extension_step",
    "formula_elements", "comp_group", "uni_test", "species")
)
}
\arguments{
\item{para_file}{'para_list.txt'}

\item{path}{'.'}

\item{showed_para}{c('package_version', 'ms1_file', 'ms2_file', 'sample_info_file',
'lib', 'polarity', 'instrument', 'column', 'ce', 'method_lc',
is_check_data', 'is_anno_initial_seed', 'is_anno_mrn', 'is_credential'
mz_tol', 'extension_step', 'formula_elements', 'comp_group', 'uni_test', 'species')}
}
\description{
generateParaList4Report
}
\author{
Zhiwei Zhou
}
