% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{generatePeakGroup}
\alias{generatePeakGroup}
\title{generatePeakGroup}
\usage{
generatePeakGroup(
  peak_table,
  base_peak_name,
  base_peak_adduct,
  base_peak_seed,
  tol_rt = 3,
  cutoff_ssc = 0
)
}
\arguments{
\item{peak_table}{peak table (row--peaks; column--samples). First 3 column should be named as "name", "mz", "rt"}

\item{base_peak_name}{name of base peak}

\item{base_peak_adduct}{adduct format of base peak}

\item{base_peak_seed}{seed of base peak}

\item{tol_rt}{rt tolerance for peak grouping; Default: 3 (second)}

\item{cutoff_ssc}{sample-sample correlation; Default: 0}
}
\description{
generatePeakGroup
}
\examples{
peak_table <- read.csv(system.file("extdata", "peak_table_200STD_neg_200805.csv", package="MetDNA2"), stringsAsFactors = F)
test <- generatePeakGroup(peak_table = peak_table,
                          base_peak_name = 'M482T929',
                          base_peak_adduct = '[M-H]-',
                          tol_rt = 3,
                          cutoff_ssc = 0)
}
\author{
Zhiwei Zhou
}
