% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Report.R
\name{generateRecursiveSummary4Report}
\alias{generateRecursiveSummary4Report}
\title{generateRecursiveSummary4Report}
\usage{
generateRecursiveSummary4Report(result_recursive, cpd_emrn)
}
\description{
generateRecursiveSummary4Report
}
\author{
Zhiwei Zhou
}
