% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMerge.R
\name{generateResultTable4AllAnnotation}
\alias{generateResultTable4AllAnnotation}
\title{generateResultTable4AllAnnotation}
\usage{
generateResultTable4AllAnnotation(
  annot_all,
  path = ".",
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  thread = 4,
  direction = c("forward", "reverse"),
  is_cred_pg_filter = TRUE,
  is_cred_formula_filter = TRUE,
  tolerance_rt_range = 30,
  test_evaluation = c("No", "200STD", "46STD")
)
}
\arguments{
\item{path}{'.'}

\item{thread}{Default: 4}
}
\description{
generate result table combining initial seed, recursive annotation, and annotation credential
}
\author{
Zhiwei Zhou
}
