% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{getAnnotationResult}
\alias{getAnnotationResult}
\title{getAnnotationResult}
\usage{
getAnnotationResult(
  tags2,
  tags.result2,
  metabolite,
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  candidate.num = 3000,
  score.cutoff = 0.4,
  ...
)
}
\description{
getAnnotationResult
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
