% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{getCcsRange}
\alias{getCcsRange}
\title{getCcsRange}
\usage{
getCcsRange(data, rel_dev)
}
\arguments{
\item{data}{Numeric.}

\item{rel_dev}{Numeric. Relative CCS deviation tolerance in percent %.}
}
\value{
a table of range. column 1: minimum value; column 2: maxmium value.
}
\description{
getCcsRange
}
\author{
Zhiwei Zhou
}
