% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{getLinerScore}
\alias{getLinerScore}
\title{getLinerScore}
\usage{
getLinerScore(delta, tolerance)
}
\arguments{
\item{delta}{Numeric. The difference between experiment value and library value.}

\item{tolerance}{The maxmium tolerance}
}
\description{
calculate liner score
}
\author{
Zhiwei Zhou
}
