% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Tools.R
\name{getPostfix}
\alias{getPostfix}
\title{getPostfix}
\usage{
getPostfix(x)
}
\arguments{
\item{x}{file_name}
}
\description{
get the postfix of file
}
\examples{
getPostfix('abc.mgf')
}
\author{
Zhiwei Zhou
}
