% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{getRtRange}
\alias{getRtRange}
\title{getRtRange}
\usage{
getRtRange(
  data,
  abs_rt_dev,
  rel_rt_dev,
  rt_threshold = 500,
  is_combined = FALSE
)
}
\arguments{
\item{data}{Numeric.}

\item{abs_rt_dev}{Numeric. Absolute rt deviation tolerance in seconds.}

\item{rel_rt_dev}{Numeric. Relative rt deviation tolerance in percent %.}
}
\value{
a table of range. column 1: minimum value; column 2: maxmium value.
}
\description{
getRtRange
}
\author{
Zhiwei Zhou
}
