% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{getTrapezoidalScore}
\alias{getTrapezoidalScore}
\title{getTrapezoidalScore}
\usage{
getTrapezoidalScore(delta, pf_range, range)
}
\arguments{
\item{delta}{Numeric. The difference between experiment value and library value.}

\item{pf_range}{Numeric. Penalty free range.}

\item{range}{Numeric. The maxmium tolerance.}
}
\description{
Trapezodial scoring function
}
\author{
Zhiwei Zhou
}
