% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{groupRT}
\alias{groupRT}
\title{groupRT}
\usage{
groupRT(rt, rt.tol = 3)
}
\arguments{
\item{rt}{The RTs of peaks.}

\item{rt.tol}{The tolerance of RT.}
}
\value{
The index of peaks in each group.
}
\description{
Group peaks according to RT.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
