% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{groupRT2}
\alias{groupRT2}
\title{groupRT2}
\usage{
groupRT2(rt_vec, rt_tol = 3)
}
\arguments{
\item{rt}{The RTs of peaks.}

\item{rt.tol}{The tolerance of RT.}
}
\value{
The index of peaks in each group.
}
\description{
Group peaks according to RT with recursive function.
}
\author{
Zhiwei Zhou
}
