% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{inteMs2}
\alias{inteMs2}
\title{inteMs2}
\usage{
inteMs2(
  ms2_data,
  ms2_type = c("mgf", "msp", "cef", "mzXML"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "WatersQTOF", "WatersTWIMMS", "AgilentDTIMMS", "BrukerTIMS"),
  metdna_version = c("version1", "version2"),
  is_include_precursor = TRUE,
  is_deisotope = FALSE,
  int_ms2_min_abs = 0,
  int_ms2_min_relative = 0.01,
  ppm_precursor_filter = 20,
  mz_range_ms2 = NULL,
  ...
)
}
\arguments{
\item{ms2_data}{ms2 data}

\item{is_include_precursor}{remove precursor peaks in raw spectra or not. Default: FALSE}

\item{is_deisotope}{remove isotope peaks in raw spectra or not. Default: TRUE}

\item{int_ms2_min_abs}{the minmium intensity of ms2 fragment. Default: 10}

\item{int_ms2_min_relative}{the minmium relative of ms2 fragment. Default: 0.01 ((1%))}

\item{ppm_precursor_filter}{the m/z tolerance to rexmove precursor ion in spectra. Default: 20}

\item{mz_range_ms2}{the range of ms2 data.}

\item{extract_ms2_file}{extract ms2 file. Default: NULL}
}
\description{
Purify and integrate multiple MS/MS files
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
