% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyInterpretation.R
\name{interpretBiology}
\alias{interpretBiology}
\title{intepretBiology}
\usage{
interpretBiology(
  sample_file_pos = "data.csv",
  sample_info_file_pos = "sample.info.csv",
  table_identification_pair_file_pos = "table3_identification_pair.csv",
  sample_file_neg = "data.csv",
  sample_info_file_neg = "sample.info.csv",
  table_identification_pair_file_neg = "table3_identification_pair.csv",
  path = ".",
  polarity = c("positive", "negative", "both"),
  metdna_version = c("version2", "version1"),
  comp_group = c("W30", "W03"),
  uni_test = c("t", "wilcox", "anova"),
  correct_p = TRUE,
  p_cutoff = 0.05,
  fc_cutoff = 1.25,
  species = c("hsa", "mmu", "rno", "bta", "gga", "dre", "dme", "cel", "sce", "osa",
    "ath", "smm", "pfa", "tbr", "eco", "bsu", "ppu", "sau", "tma", "syf", "mln"),
  quanti_pathway_method = c("mean", "sum", "median"),
  organze_basis = c("kegg_id", "all"),
  extension_step = c("0", "1", "2", "3", "4", "5", "6", "7", "8"),
  ...
)
}
\arguments{
\item{path}{'.'}

\item{comp_group}{the group for comparison, which needed included in sample information}

\item{uni_test}{univariate test, including "t","wilcox","anova". Default: 't'}

\item{correct_p}{whether correct P value}

\item{species}{"hsa", "mmu", "rno", "bta", "gga", "dre", "dme", "cel", "sce", "osa", "ath", "smm", "pfa", "tbr", "eco", "bsu", "ppu", "sau", "tma", "syf", "mln"}

\item{quanti_pathway_method}{the mehtod of pathway quantification, including 'mean', 'sum', 'median'. Default: 'mean'}

\item{extension_step}{c('0', '1', '2', '3', '4', '5', '6', '7', '8')}
}
\description{
a wrapper for MetDNA2
}
\examples{

}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
