% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{loadSpecDB}
\alias{loadSpecDB}
\title{loadSpecDB}
\usage{
loadSpecDB(
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "WatersQTOF", "WatersTWIMMS", "AgilentDTIMMS", "BrukerTIMS"),
  column = c("hilic", "rp"),
  method_lc = c("Amide12min", "Amide23min", "MetlinRP", "Other", "RP12min"),
  ce = c("30", "10", "20", "35,15", "40", "50", "NCE10", "NCE20", "NCE30", "NCE40",
    "NCE50", "SCE20_30_40\%", "SNCE20_30_40\%"),
  polarity = c("positive", "negative"),
  adduct_list = NULL,
  file_rt_ref = "RT_recalibration_table.csv",
  is_rt_calibration = TRUE,
  path = "."
)
}
\arguments{
\item{lib}{database name, 'zhuMetLib', 'zhuMetLib_orbitrap', 'fiehnHilicLib'. Default: 'zhuMetLib'}

\item{instrument}{"SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS". Default: "SciexTripleTOF"}

\item{column}{'hilic', 'rp'. Default: 'hilic'}

\item{method_lc}{'Amide12min', Amide23min', 'Other', 'MetlinRP'}

\item{ce}{"10", "20", "30", "35,15", "40", "50"; Default: '30'}

\item{polarity}{'positive' or 'negative'. Default: 'positive'}

\item{adduct_list}{NULL}

\item{is_rt_calibration}{TRUE}

\item{path}{'.'}
}
\description{
loadSpecDB
}
\author{
Zhiwei Zhou
}
