% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{matchMs1WithSpecLib}
\alias{matchMs1WithSpecLib}
\title{matchMs1WithSpecLib}
\usage{
matchMs1WithSpecLib(
  ms1_data,
  lib_meta,
  mz_tol = 25,
  mz_ppm_thr = 400,
  pf_rt_range = 0,
  tolerance_rt_range = 30,
  pf_ccs_range = 0,
  tolerance_ccs_range = 4,
  is_filter = TRUE,
  is_rt_score = FALSE,
  is_ccs_score = TRUE,
  is_check_cation = TRUE
)
}
\arguments{
\item{ms1_data}{ms1 data, including info and subject.}

\item{lib_meta}{the meta information of library}

\item{mz_tol}{m/z tolerance. Default: 10ppm}

\item{pf_ccs_range}{penalty-free CCS range. Default: 0}

\item{tolerance_ccs_range}{maxmium CCS tolerance. Default: 4%}

\item{is_ccs_score}{logical vector. Default: TRUE}
}
\description{
matchMs1WithSpecLib
}
\author{
Zhiwei Zhou
}
