% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{matchMs2WithNeighbor}
\alias{matchMs2WithNeighbor}
\title{matchMs2WithNeighbor}
\usage{
matchMs2WithNeighbor(
  peak_int,
  tags2,
  seed_idx,
  ms2,
  round,
  metabolite,
  metabolite_ccs,
  metabolic_network,
  max_isotope = 4,
  polarity = c("positive", "negative"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  rt_tol1 = 3,
  rt_tol2 = 30,
  mz_tol = 25,
  mz_ppm_thr = 400,
  cor_tol = 0,
  int_tol = 500,
  ccs_tol = 4,
  dp_tol = 0.5,
  matched_frag_tol = 1,
  whether_link_frag = FALSE,
  max_step = 3,
  mz_tol_ms2 = 25,
  remove_index = NULL,
  iso_annotation = TRUE,
  add_annotation = TRUE,
  met_annotation = TRUE,
  threads = 3,
  whether_use_predRT = TRUE,
  adduct_table,
  ...
)
}
\arguments{
\item{peak_int}{The median/mean intensity of peaks.}

\item{tags2}{The tags2 data of all peaks.}

\item{seed_idx}{The index of seeds.}

\item{ms2}{The ms2 data of peak table.}

\item{round}{The round of annotation.}

\item{metabolite}{The kegg compound database.}

\item{metabolic_network}{kegg.rpair2}

\item{max_isotope}{The number of isotope peaks.}

\item{polarity}{The polarity.}

\item{rt_tol1}{The RT tolerance for isotope and adduct annotation. (second)}

\item{rt_tol2}{The RT tolerance for metabolite annotation (\%).}

\item{mz_tol}{The mz tolerance.}

\item{cor_tol}{The cor tolerance.}

\item{int_tol}{The intensity ratio tolerance.}

\item{dp_tol}{The tolerance of dot product.}

\item{max_step}{The max number of reaction step.}

\item{remove_index}{The index of peaks which you dont want to annotate.}

\item{iso_annotation}{Isotope annotation or not.}

\item{add_annotation}{Adduct annotation or not.}

\item{met_annotation}{Metabolite annotation or not.}

\item{threads}{The number of threads.}

\item{adduct_table}{Adduct table.}
}
\value{
The tags data with annotation result.
}
\description{
Annotate peak table from seeds.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
