% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{matchMs2WithSpecLib}
\alias{matchMs2WithSpecLib}
\title{matchMs2WithSpecLib}
\usage{
matchMs2WithSpecLib(
  ms1_result,
  exp_spec,
  lib_meta,
  lib_spec,
  metdna_version = c("version2", "version1"),
  mz_tol_ms2 = 35,
  dp_cutoff = 0.8,
  matched_frag_cutoff = 1,
  path = ".",
  is_include_precursor = TRUE,
  direction = c("reverse", "forward"),
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  ...
)
}
\arguments{
\item{ms1_result}{ms1 result}

\item{exp_spec}{correspond msms data}

\item{lib_meta}{library of meta information}

\item{lib_spec}{library of spectrum}

\item{mz_tol_ms2}{Default: 35 ppm}

\item{dp_cutoff}{Default: 0.8}

\item{matched_frag_cutoff}{Default: 1}

\item{direction}{'forward', 'reverse'}

\item{scoring_approach}{'dp', 'bonanza', 'hybrid', 'gnps'. Default: 'dp'}
}
\description{
matchMs2WithSpecLib
}
\author{
Zhiwei Zhou
}
