% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMerge.R
\name{mergeRecursiveAnnotation}
\alias{mergeRecursiveAnnotation}
\title{mergeRecursiveAnnotation}
\usage{
mergeRecursiveAnnotation(path = ".", use_redun_rm_result = FALSE, ...)
}
\arguments{
\item{path}{'.'}
}
\description{
mergeRecursiveAnnotation
}
\author{
Zhiwei Zhou
}
