% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2.R
\name{MetDNA2}
\alias{MetDNA2}
\title{MetDNA2}
\usage{
MetDNA2(
  ms1_file = "data.csv",
  ms2_file = NULL,
  sample_info_file = "sample.info.csv",
  metdna_version = c("version2", "version1"),
  ms2_type = c("mgf", "mzXML", "msp", "cef"),
  path = ".",
  thread = 4,
  is_check_data = TRUE,
  is_anno_initial_seed = TRUE,
  is_anno_mrn = TRUE,
  is_credential = TRUE,
  is_bio_interpret = TRUE,
  is_exported_report = TRUE,
  is_cred_pg_filter = TRUE,
  is_cred_formula_filter = FALSE,
  is_rm_intermediate_data = FALSE,
  lib = c("zhuMetLib", "zhuMetLib_orbitrap", "fiehnHilicLib", "zhuRPLib"),
  polarity = c("positive", "negative"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "WatersQTOF",
    "ThermoOrbitrap", "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersTWIMMS"),
  column = c("hilic", "rp"),
  ce = c("30", "10", "20", "35,15", "40", "50", "NCE10", "NCE20", "NCE30", "NCE40",
    "NCE50", "SCE20_30_40\%", "SNCE20_30_40\%"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "zhulabRP"),
  is_rt_calibration = TRUE,
  mz_tol = 25,
  mz_ppm_thr = 400,
  tolerance_rt_range = 30,
  tolerance_ccs_range = 4,
  int_ms2_min_abs = 50,
  dp_cutoff = 0.8,
  direction = c("reverse", "forward"),
  mz_tol_combine_ms1_ms2 = 25,
  rt_tol_combine_ms1_ms2 = 10,
  is_plot_ms2 = TRUE,
  adduct_limit_specLib = "all",
  test_force_filtering_rt = NULL,
  extension_step = c("0", "1", "2", "3", "4", "5", "6", "7", "8"),
  max_isotope = 4,
  rt_tol1 = 3,
  rt_tol2 = 30,
  ccs_tol = 4,
  cor_tol = 0,
  int_tol = 500,
  dp_tol = 0.5,
  max_step = 3,
  score_cutoff = 0,
  remain = FALSE,
  remain_per = 0.5,
  seed_neighbor_match_plot = TRUE,
  candidate_num = 5,
  scoring_approach_recursive = c("dp", "bonanza", "hybrid", "gnps"),
  matched_frag_cutoff = 1,
  whether_link_frag = FALSE,
  use_pretrained_model = FALSE,
  use_redun_rm_result = TRUE,
 
    dir_GenForm = "/home/zhouzw/software/genform-code-r8-trunk/src/GenForm/GenForm/bin/Release",
  is_pred_formula_all = FALSE,
  formula_mz_tol_ms1 = 25,
  formula_mz_tol_ms2 = 25,
  formula_elements = "CHNOPS",
  formula_candidate_num = 3,
  platform = "linux",
  test_rm_extra_anno_from_ini_seed = FALSE,
  is_plot_pseudo_MS1 = TRUE,
  remove_conflict_seed_from_final_table = FALSE,
  comp_group = c("W03", "W30"),
  uni_test = c("t", "wilcox", "anova"),
  correct_p = FALSE,
  p_cutoff = 0.05,
  fc_cutoff = 1,
  species = c("hsa", "mmu", "rno", "bta", "gga", "dre", "dme", "cel", "sce", "osa",
    "ath", "smm", "pfa", "tbr", "eco", "bsu", "ppu", "sau", "tma", "syf", "mln"),
  quanti_pathway_method = c("mean", "sum", "median"),
  test_evaluation = c("No", "200STD", "46STD")
)
}
\arguments{
\item{ms1_file}{The name of ms1 data}

\item{ms2_file}{The names of ms2 data}

\item{sample_info_file}{'sample.info.csv'}

\item{metdna_version}{MetDNA parameter 'version1' and 'version2'}

\item{ms2_type}{'mgf', 'mzXML', 'msp', 'cef' are supported. Default: 'mgf'}

\item{path}{Default: '.'}

\item{thread}{The number of threads. Default: 4}

\item{is_check_data}{Whether check data quality. Default: TRUE}

\item{is_anno_initial_seed}{Whether annotate initial seeds with spectral DB. Default: TRUE}

\item{is_anno_mrn}{Whether perform recursive annotation. Default: TRUE}

\item{is_credential}{Whether perform annotation credential. Default: TRUE}

\item{is_cred_formula_filter}{Whether filter error annotations by formula prediction in final table. Default: FALSE}

\item{lib}{Default: 'zhuMetLib'}

\item{polarity}{'positive', 'negative'. Default: 'positive'}

\item{instrument}{The instrument you used to acquire data. "SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
'ThermoExploris', "AgilentDTIMMS", "BrukerTIMS", 'WatersQTOF','WatersTWIMMS'. Default: "SciexTripleTOF"}

\item{column}{"hilic", "rp". Default: 'hilic'}

\item{ce}{"30", "10", "20", "35,15", "40", "50". Default: "30"}

\item{method_lc}{'Amide12min', 'Amide23min', 'Other', 'MetlinRP'. Default: "Amide12min"}

\item{is_rt_calibration}{Only valid with 'Amide12min' and 'Amide23min' methods. Default: TRUE.}

\item{mz_tol}{The tolerance of ms1 match. Default: 25 ppm}

\item{tolerance_rt_range}{The maxmium rt tolerance. Default: 30 s}

\item{dp_cutoff}{MS2 score cutoff. Default: 0.8}

\item{direction}{Reserve which direct dot product. 'reverse', 'forward'. Default: 'reverse'}

\item{is_plot_ms2}{Whether plot initial-seed ms2 match plot. Default: TRUE}

\item{extension_step}{The used extended MRN with n step reaction. Steps: '0', '1', '2', '3', '4', '5', '6', '7', '8'}

\item{max_isotope}{Maximum considerable isotope. Default: 4}

\item{rt_tol1}{RT tolerance of adductAnnotation and isotopeAnnotation. Default: 3s}

\item{rt_tol2}{RT tolerance of metAnnotation. Default: 30%}

\item{cor_tol}{Default: 0}

\item{int_tol}{isotope intensity tolerance. Defautl: 500}

\item{dp_tol}{0.5}

\item{max_step}{the maximum step of reaction pair during recursive annotation}

\item{score_cutoff}{Total score cutoff. Default: 0}

\item{remain}{FALSE}

\item{remain_per}{0.5}

\item{seed_neighbor_match_plot}{TRUE}

\item{candidate_num}{reserve top N candidate number in recursive annotation. Default: 3}

\item{is_pred_formula_all}{whether predict formula for all peak group (TRUE) or concised peak groups (FALSE). Default: FALSE}

\item{formula_mz_tol_ms1}{Default: 25 ppm}

\item{formula_mz_tol_ms2}{Default: 25 ppm}

\item{formula_elements}{Default: "CHNOPS"}

\item{formula_candidate_num}{Reserve top 3 formula. Default: 3}

\item{platform}{'windows', 'linux'. Default: 'linux'}

\item{is_cred_pg_filer}{Whether filter error annotations by peak group annotaion in final table. Default: FALSE}
}
\description{
Metabolite annotation and dysregulated network analysis.
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
