% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{modifyOutputResult}
\alias{modifyOutputResult}
\title{modifyOutputResult}
\usage{
modifyOutputResult(id_merge, test_rm_extra_anno_from_ini_seed = FALSE)
}
\description{
output id_merge result as level1, level2, level3.1, level3.2, level4
}
\author{
Zhiwei Zhou
}
