% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Class.R
\docType{class}
\name{PeakInfo-class}
\alias{PeakInfo-class}
\title{PeakInfo}
\description{
The PeakInfo class is the S4 class for Peak annotation information.
}
\section{Slots}{

\describe{
\item{\code{name}}{Peak name.}

\item{\code{mz}}{Peak m/z.}

\item{\code{rt}}{Peak rentention time (RT).}

\item{\code{ms2}}{MS/MS spectrum of peak.}

\item{\code{annotation}}{Annotation information of peak. It is a list contains different annotation result.}
}}

