% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotIdMs2}
\alias{plotIdMs2}
\title{plotIdMs2}
\usage{
plotIdMs2(obj_ms2 = NULL, obj_spec = NULL)
}
\arguments{
\item{obj_ms2}{object of SpectraData}
}
\description{
generate ms2 mirror plot
}
\author{
Zhiwei Zhou
}
