% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotIdShiftMs2}
\alias{plotIdShiftMs2}
\title{plotIdShiftMs2}
\usage{
plotIdShiftMs2(obj_ms2 = NULL, obj_spec_frag_match, obj_spec_frag_nl)
}
\arguments{
\item{obj_ms2}{object of SpectraData}
}
\description{
generate ms2 mirror plot for shift match (bonanza, hybrid, gnps)
}
\author{
Zhiwei Zhou
}
