% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotNeighbor}
\alias{plotNeighbor}
\title{plotNeighbor}
\usage{
plotNeighbor(
  tags2,
  path = ".",
  ms2 = ms2,
  kegg.compound = kegg.compound,
  matchParam = matchParam,
  ...
)
}
\arguments{
\item{path}{'.'}

\item{ms2}{ms2 RData}
}
\description{
export neighbor match plot during recursive annotation
}
\author{
Zhiwei Zhou
}
