% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPathwayBar}
\alias{plotPathwayBar}
\title{plotPathwayBar}
\usage{
plotPathwayBar(result_pathway_enrichment, p_cutoff = 0.05, p_adjust = TRUE)
}
\description{
plotPathwayBar
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
