% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPathwayBoxplot}
\alias{plotPathwayBoxplot}
\title{plotPathwayBoxplot}
\usage{
plotPathwayBoxplot(
  quanti_pathway_data,
  sample_info,
  comp_group = c("W30", "W03"),
  path = "."
)
}
\arguments{
\item{path}{'.'}
}
\description{
plotPathwayBoxplot
}
