% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPathwayHeatmap}
\alias{plotPathwayHeatmap}
\title{plotPathwayHeatmap}
\usage{
plotPathwayHeatmap(
  quanti_pathway_data,
  sample_info,
  comp_group = c("W30", "W03")
)
}
\description{
plotPathwayHeatmap
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
