% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPathwayHeatmapMet}
\alias{plotPathwayHeatmapMet}
\title{plotPathwayHeatmapMet}
\usage{
plotPathwayHeatmapMet(
  quanti_met_data,
  quanti_pathway_data,
  sample_info,
  lib_pathway,
  comp_group = c("W30", "W03"),
  species = c("hsa", "mmu", "rno", "bta", "gga", "dre", "dme", "cel", "sce", "osa",
    "ath", "smm", "pfa", "tbr", "eco", "bsu", "ppu", "sau", "tma", "syf", "mln"),
  path = ".",
  ...
)
}
\arguments{
\item{path}{'.'}
}
\description{
plotPathwayHeatmapMet
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
