% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPathwayScatter}
\alias{plotPathwayScatter}
\title{plotPathwayScatter}
\usage{
plotPathwayScatter(result_pathway_enrichment, p_cutoff = 0.05, p_adjust = TRUE)
}
\description{
plotPathwayScatter
}
\author{
Zhiwei Zhou
}
