% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotPseudoMs1Spec}
\alias{plotPseudoMs1Spec}
\title{plotPseudoMs1Spec}
\usage{
plotPseudoMs1Spec(peak_group)
}
\description{
generate Pseudo MS1 spectrum for annotated peak group
}
\examples{
load(system.file("tempdata", "list_peak_group_annotation_200805.RData", package="MetDNA2"))
list_peak_group_annotation[[35]] \%>\% plotPseudoMs1Spec()
}
\author{
Zhiwei Zhou
}
