% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{plotRtCalibration}
\alias{plotRtCalibration}
\title{plotRtCalibration}
\usage{
plotRtCalibration(result_rt_calibration, rt_recalibration_model, path = ".")
}
\arguments{
\item{path}{'.'}
}
\description{
plotRtCalibration
}
\author{
Zhiwei Zhou
}
