% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{predictRT}
\alias{predictRT}
\title{predictRT}
\usage{
predictRT(
  data,
  prefer_adduct = c("all", "M+Na", "M+H", "M+NH4", "M-H"),
  threads = 3,
  md_inHouse_cpd,
  md_kegg_cpd,
  use_default_md = TRUE,
  column = c("hilic", "rp"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min"),
  metdna_version = c("version2", "version1"),
  use_pretrained_model = FALSE
)
}
\arguments{
\item{data}{data From readAnnotation.}

\item{threads}{How many threads do you want to use? Default is the number of your PC threads - 3.}

\item{prefer.adduct}{The reliable adducts in this LC system. Default is all.}

\item{inHouse.compound.md}{The molecular descriptors of inhouse compounds.}

\item{kegg.compound.md}{The molecular descriptors of kegg compounds.}
}
\value{
The predicted in house RT and KEGG RT.
}
\description{
Predict RTs of kegg metabolites using MS2 matched metabolites.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
