% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{prepareMetAnalyzer}
\alias{prepareMetAnalyzer}
\title{prepareMetAnalyzer}
\usage{
prepareMetAnalyzer(
  peak_table_file = "idresults.csv",
  ms2_file = "spectra.msp",
  sample_info_file = "sample.info.csv",
  path_dir = ".",
  lib = c("zhulib", "zhumetlib"),
  match_method = c("forward", "reverse"),
  polarity = c("positive", "negative"),
  verison = c("1", "2"),
  ...
)
}
\description{
generate files (MetAnalyzer result) for annotation credential
}
\author{
Zhiwei Zhou
}
