% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{prepareMetDNA}
\alias{prepareMetDNA}
\title{prepareMetDNA}
\usage{
prepareMetDNA(
  peak_table_file = "data.csv",
  sample_info_file = "sample.info.csv",
  path_dir = ".",
  polarity = c("positive", "negative")
)
}
\description{
generate files for annotation credential
}
\author{
Zhiwei Zhou
}
