% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{prepareMetDNA2}
\alias{prepareMetDNA2}
\title{prepareMetDNA2}
\usage{
prepareMetDNA2(
  peak_table_file = "data.csv",
  sample_info_file = "sample.info.csv",
  direction = c("reverse", "forward"),
  path_dir = ".",
  polarity = c("positive", "negative"),
  ...
)
}
\description{
generate files for annotation credential
}
\author{
Zhiwei Zhou
}
