% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{purifyMs2}
\alias{purifyMs2}
\title{purifyMs2}
\usage{
purifyMs2(
  spec,
  mz_precursor,
  is_include_precursor = FALSE,
  is_deisotope = FALSE,
  is_remove_ring_effect = TRUE,
  mz_range_ms2 = NULL,
  int_ms2_min_abs = 10,
  int_ms2_min_relative = 0.03,
  ppm_precursor_filter = 20
)
}
\arguments{
\item{spec}{the spectra matrix. column1: mz; column2: intensity}

\item{mz_precursor}{the precursor m/z. Numeric.}

\item{is_include_precursor}{remove precursor peak in raw spectra or not_ Default: False}

\item{mz_range_ms2}{the range of m/z. Default: NULL}

\item{int_ms2_min_abs}{the minmium intensity of ms2 fragment_ Default: 10}

\item{int_ms2_min_relative}{the minmium relative of ms2 fragment_ Default: 0.03 ((3%))}

\item{ppm_precursor_filter}{the m/z tolerance to remove precursor ion in spectra_ Default: 20}
}
\description{
purifyMs2
}
\author{
Zhiwei Zhou
}
