% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{rankGenFormFormula}
\alias{rankGenFormFormula}
\title{rankGenFormFormula}
\usage{
rankGenFormFormula(file_genform = "ms2_file_GenForm_sorted.txt")
}
\arguments{
\item{file_genform}{the file name of GenForm result}
}
\description{
rank predicted formula with known formula DB
}
\examples{
rankGenFormFormula(file_genform = 'I:/00_projects/03_MetDNA2/00_data/20200810_conbine_Genform_to_feature_credential/example/result/SinapinicAcidMsMs_GenForm_sorted.txt')
}
\author{
Zhiwei Zhou
}
