% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImporting.R
\name{readCEF}
\alias{readCEF}
\title{readCEF}
\usage{
readCEF(file)
}
\arguments{
\item{file}{path of the CEF file}
}
\description{
Read a CEF file and return a list of spectra for all feature
 with feature information
}
