% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{readInitialAnnotation}
\alias{readInitialAnnotation}
\title{readInitialAnnotation}
\usage{
readInitialAnnotation(
  data = "ms2_match_annotation_result.csv",
  path = ".",
  rt_filter = FALSE,
  rt_tol = 30,
  direction = c("reverse", "forward"),
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS"),
  inHouse_compound,
  ...
)
}
\arguments{
\item{data}{The file name (csv) from ms2Annotation}

\item{path}{The work directory}

\item{rt_filter}{Filter annotation according or not.}

\item{rt_tol}{The tolerance of RT (\%).}

\item{direction}{= c('reverse', 'forward')}

\item{inHouse_compound}{in house compound database.}
}
\description{
Read the data from ms2Annotation and seperate it to tags and sample
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
