% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{readMs1}
\alias{readMs1}
\title{readMs1}
\usage{
readMs1(
  filename,
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "WatersQTOF", "WatersTWIMMS", "AgilentDTIMMS", "BrukerTIMS"),
  path = "."
)
}
\arguments{
\item{filename}{The name of csv file.}

\item{instrument}{"SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", 'ThermoExploris', "AgilentDTIMMS", "BrukerTIMS", 'WatersQTOF','WatersTWIMMS'}

\item{path}{'.'}
}
\value{
A list. info: a table of feature information, column 1_feature name, column 2_mz, column3_rt in seconds, column4_ccs. subject: a table of peak area.
}
\description{
readMs1
}
\author{
Zhiwei Zhou
}
