% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImporting.R
\name{readMs2}
\alias{readMs2}
\title{readMs2}
\usage{
readMs2(ms2_file, ms2_type = c("mgf", "cef", "msp", "mzXML"), ...)
}
\arguments{
\item{ms2_file}{MS/MS file names}

\item{ms2_type}{'mgf', 'cef', 'msp', 'MzXML'}
}
\description{
readMs2
}
\author{
Zhiwei Zhou
}
