% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{refineAnnotation}
\alias{refineAnnotation}
\title{refineAnnotation}
\usage{
refineAnnotation(
  annotation_initial,
  list_peak_group_annotation_concised,
  list_peak_group_formula,
  path_dir = "."
)
}
\arguments{
\item{path_dir}{Default: '.'}
}
\description{
refine annotation credential with semi-targeted clustering and formula prediction
}
\author{
Zhiwei Zhou
}
