% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{refineAnnotationFeatureClustering}
\alias{refineAnnotationFeatureClustering}
\title{refineAnnotationFeatureClustering}
\usage{
refineAnnotationFeatureClustering(
  annotation_initial,
  list_peak_group_annotation_concised
)
}
\description{
refineAnnotationFeatureClustering
}
\author{
Zhiwei Zhou
}
