% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredential.R
\name{refineAnnotationFormula}
\alias{refineAnnotationFormula}
\title{refineAnnotationFormula}
\usage{
refineAnnotationFormula(
  annotation_initial,
  list_peak_group_formula,
  num_formula_candidate = 3,
  path_dir,
  ...
)
}
\description{
refineAnnotationFormula
}
\author{
Zhiwei Zhou
}
