% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{removeIsotopes}
\alias{removeIsotopes}
\title{removeIsotopes}
\usage{
removeIsotopes(spec)
}
\arguments{
\item{spec}{the matrix spectra. 1st column: "mz", 2nd column: "intensity"}
}
\description{
removeIsotopes
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
