% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{removeRedundancy}
\alias{removeRedundancy}
\title{removeRedundancy}
\usage{
removeRedundancy(result, polarity = c("positive", "negative"), path = ".")
}
\arguments{
\item{result}{The result from metA.}

\item{path}{The directory.}
}
\value{
A result after removing redundancy.
}
\description{
Remove annotation and peak redundancy.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
