% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Result.R
\name{removeReplicateSeedAnnotationFromFinalTable}
\alias{removeReplicateSeedAnnotationFromFinalTable}
\title{removeReplicateSeedAnnotationFromFinalTable}
\usage{
removeReplicateSeedAnnotationFromFinalTable(
  dir_path,
  instrument = c("SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap",
    "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS", "WatersQTOF", "WatersTWIMMS")
)
}
\arguments{
\item{dir_path}{project path}
}
\description{
remove replicated metabolite annotations between seed annotation and recursive annotation
}
\author{
Zhiwei Zhou
}
