% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{removeRingEffect}
\alias{removeRingEffect}
\title{removeRingEffect}
\usage{
removeRingEffect(spec, mz_diff_thr = 0.3, int_rel_thr = 0.2)
}
\arguments{
\item{spec}{2-column table. column 1: mz; column 2: intensity}

\item{mz_diff_thr}{the threold of mz difference. Default: 0.3}

\item{int_rel_thr}{the threold of relative intensity. Default: 0.2}
}
\description{
removeRingEffect
}
\author{
Yandong Yin; Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
