% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Tools.R
\name{retieveUpperPath}
\alias{retieveUpperPath}
\title{'retieveUpperPath'}
\usage{
retieveUpperPath(
  path = "/home/zhouzw/Data_processing/20210626_metdna2_merge_pos_neg/BOTH/"
)
}
\description{
'retieveUpperPath'
}
\examples{
retieveUpperPath("/home/zhouzw/Data_processing/20210626_metdna2_merge_pos_neg/BOTH/")
}
\author{
Zhiwei Zhou
}
