% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{runGenForm}
\alias{runGenForm}
\title{runGenForm}
\usage{
runGenForm(
  ms_file,
  msms_file,
  ion,
  dir_GenForm,
  dir_result,
  ppm = 5,
  acc = 15,
  elements = "CHNOPS",
  ff = "",
  dbe = TRUE,
  oei = TRUE,
  mz = 0,
  exist = TRUE,
  an_loss = TRUE,
  an_loss_limit = 500,
  sort = TRUE,
  cleanMSMS = FALSE,
  platform = c("windows", "linux")
)
}
\arguments{
\item{ms_file}{File name of a plain text, two column file (e.g. \code{MS.txt}) containing
space-separated mz and intensity values of the MS1 spectrum (full scan). The most intense
peak is taken as "m" unless this is defined with \code{"mz"}.}

\item{msms_file}{A plain text, two column file (e.g. \code{MSMS.txt}) containing
space-separated mz and intensity values of the tandem (MS2) spectrum, i.e. the fragments.
While GenForm can be run without MSMS, this function requires its use at this stage.}

\item{ion}{Ion setting to use. Options: \code{c("-e","+e","-H","+H","+Na")}.}

\item{dir_GenForm}{Directory containing \code{GenForm.exe}.}

\item{dir_result}{Directory to save the GenForm results. Existing files will be overwritten.}

\item{ppm}{Accuracy setting for MS1 (default \code{5} ppm).}

\item{acc}{Accuracy setting for MS2 (default \code{15} ppm).}

\item{elements}{Selection of elements to consider. Default \code{"CHNOPS"}, implemented
are \code{"CHBrClFINOPSSi"}.}

\item{ff}{A "fuzzy formula" input to offer more control beyond \code{elements}. This should be
in format \code{C0-6H0-20O1-3}.}

\item{dbe}{Default \code{TRUE} writes double bond equivalents to output; \code{FALSE} suppresses
this output.}

\item{oei}{Default \code{TRUE} allows odd electron ions for explaining MS/MS peaks, \code{FALSE}
means only even electron ions are considered.}

\item{mz}{If mz is >0, this is the mass GenForm will use to calculate the formulas, considering the \code{ion} setting. Default is the most intense peak in \code{ms_file}.}

\item{exist}{Default \code{TRUE} activates the existance filter (allowing only formulae for which at least one structural formula exists). \code{FALSE} is not recommended for most use cases.}

\item{an_loss}{Activate the annotation of MSMS fragments with subformulas and losses.}

\item{an_loss_limit}{Define the cut-off limit for annotating subformulas and losses. For many peaks and formulas, this becomes time consuming and produces massive files of limited use.}

\item{sort}{Default \code{FALSE} will produce default output. If \code{TRUE}, generates an additional
sorted file sorted by combined match value.}

\item{cleanMSMS}{Default \code{FALSE}. If \code{TRUE}, produces an additional cleaned MSMS file containing only peaks with subformula annotation.}
}
\value{
Several files containing various GenForm outputs.
}
\description{
This is a wrapper function for GenForm, requiring MS, MS/MS, ion settings
and directories. This function is modified from GenFormR
}
\examples{
runGenForm(ms_file = 'I:/00_projects/03_MetDNA2/00_data/20200810_conbine_Genform_to_feature_credential/example/SinapinicAcidMs.txt',
           msms_file = 'I:/00_projects/03_MetDNA2/00_data/20200810_conbine_Genform_to_feature_credential/example/SinapinicAcidMsMs.txt',
           ion = c('+H'),
           dir_GenForm = 'H:/software/GeneForm',
           dir_result = 'I:/00_projects/03_MetDNA2/00_data/20200810_conbine_Genform_to_feature_credential/example/result2',
           ppm = 10,
           elements = "CHNOPS",
           ff = "",
           dbe = TRUE,
           oei = TRUE,
           mz = 0,
           exist = TRUE,
           an_loss = TRUE,
           an_loss_limit = 500,
           sort = TRUE,
           cleanMSMS = FALSE
           )
}
\references{
GenForm on Source Forge: https://sourceforge.net/projects/genform/
}
\author{
Emma Schymanski (R wrapper, \code{\link{emma.schymanski@uni.lu}}) and Markus Meringer (GenForm); Zhiwei Zhou (Modified for MetDNA2)
}
