% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationInitialSeed.R
\name{runSpecMatch}
\alias{runSpecMatch}
\title{runSpecMatch}
\usage{
runSpecMatch(
  obj_ms2_cpd1,
  obj_ms2_cpd2,
  mz_tol_ms2 = 35,
  scoring_approach = c("dp", "bonanza", "hybrid", "gnps"),
  ...
)
}
\arguments{
\item{obj_ms2_cpd1}{experimental ms2 object}

\item{obj_ms2_cpd2}{library ms2 object}

\item{mz_tol_ms2}{Default: 35 ppm}

\item{scoring_approach}{'dp', 'bonanza', 'hybrid', 'gnps'}
}
\description{
a interphace of runing SpectraTools
}
\author{
Zhiwei Zhou
}
