% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialFormula.R
\name{searchFormulaDB}
\alias{searchFormulaDB}
\title{searchFormulaDB}
\usage{
searchFormulaDB(target_formula, score_bio_db = 0.3, score_pred_db = 0.15, ...)
}
\arguments{
\item{score_bio_db}{Default: 0.3}

\item{score_pred_db}{Default: 0.15}
}
\value{
additional score
}
\description{
search formula in known compound formula DB to plus additional score for ranking. The MsfinderFormulaDB-VS10.efd (MSFINDER DB) was used as known compound DB.
}
\examples{
searchFormulaDB(target_formula = 'C6H12O6')
}
\author{
Zhiwei Zhou
}
