% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Class.R
\name{showTags2}
\alias{showTags2}
\title{showTags2}
\usage{
showTags2(
  tags2,
  slot = c("name", "mz", "rt", "ccs", "annotation.len", "annotation.id",
    "annotation.from", "annotation.from.peak", "annotation.score", "annotation.level",
    "annotation.step", "seed.number", "annotation.type", "annotation.formula", "as.seed")
)
}
\arguments{
\item{tags2}{The tags2 data.}

\item{slot}{The inforamtion.}
}
\value{
The information of tags2 data.
}
\description{
Show detail information of tags2 data.
}
\author{
Xiaotao Shen, Zhiwei Zhou
}
