% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Tools.R
\name{splitInchiKey}
\alias{splitInchiKey}
\title{splitInchiKey}
\usage{
splitInchiKey(inchikey = "AEMRFAOFKBGASW-UHFFFAOYSA-N")
}
\value{
a data.frame with 4 columns
}
\description{
split inchikey to 3 parts
}
\author{
Zhiwei Zhou
}
