% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{statNumListPeakGroup}
\alias{statNumListPeakGroup}
\title{statNumListPeakGroup}
\usage{
statNumListPeakGroup(list_peak_group)
}
\description{
statistics number of peak groups
}
\examples{
load(system.file("tempdata", "list_peak_group_annotation_200805.RData", package="MetDNA2"))
statNumListPeakGroup(list_peak_group = list_peak_group_annotation)
}
\author{
Zhiwei Zhou
}
