% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCredentialPeakGroup.R
\name{statTypeListPeakGroup}
\alias{statTypeListPeakGroup}
\title{statTypeListPeakGroup}
\usage{
statTypeListPeakGroup(list_peak_group)
}
\description{
statTypeListPeakGroup
}
\author{
Zhiwei Zhou
}
