% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Tools.R
\name{SXTMTmatch}
\alias{SXTMTmatch}
\title{SXTMTmatch}
\usage{
SXTMTmatch(
  data1,
  data2,
  mz.tol = 25,
  rt.tol = 10,
  ccs.tol = NULL,
  rt.error.type = c("relative", "abs")
)
}
\arguments{
\item{data1}{First data for matching, first column must be mz and seconod column must be rt.}

\item{data2}{Second data for matching, first column must be mz and seconod column must be rt.}

\item{mz.tol}{mz tol for ms1 and ms2 data matching. ppm}

\item{rt.tol}{RT tol for ms1 and ms2 data matching. s or %}
}
\value{
Return a result which give the matching result of data1 and database.
}
\description{
Match two data according to mz and RT.
}
\author{
Xiaotao Shen, Zhiwei Zhou
}
