% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{trainRandomForestWithCaret}
\alias{trainRandomForestWithCaret}
\title{trainRandomForestWithCaret}
\usage{
trainRandomForestWithCaret(x)
}
\arguments{
\item{x}{the first column is RT, others are MDs}
}
\value{
RF model
}
\description{
trainRandomForestWithCaret
}
\author{
Zhiwei Zhou
}
