% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveAnnotationMRN.R
\name{trans2Matrix}
\alias{trans2Matrix}
\title{trans2Matrix}
\usage{
trans2Matrix(tags2, base = c("annotation", "peak"))
}
\arguments{
\item{tags2}{The tags2 data.}

\item{base}{Sort according to annotation or peak.}
}
\value{
The matrix of tags2 data.
}
\description{
Transform tags2 data to matrix.
}
\author{
Zhiwei Zhou, Xiaotao Shen
}
