% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDNA2Plots.R
\name{ZZWTheme}
\alias{ZZWTheme}
\title{ZZWTheme}
\usage{
ZZWTheme(type = c("common", "classic"))
}
\arguments{
\item{type:}{'common', 'classic'}
}
\description{
ggplot theme
}
\examples{
mydata <- data.frame(
 Lebal  = c("linerange1","linerange2","linerange3","linerange4","linerange5"),
 xstart = c(3.5,7,12,16,20),
 ymin   = c(2.5,6.5,3,4.5,3.8),
 ymax   = c(7.5,9.5,9,13.5,4.2),
 class  = c("A","A","A","C","C")
)
test <- ggplot(mydata) +
 geom_linerange(aes(x = xstart, ymin = ymin , ymax = ymax , colour = class) , size = 1.5)
test
mytheme <- ZZWTheme()
test + mytheme
mytheme2 <- ZZWTheme('classic')
test +
  mytheme2 +
  theme(axis.text.y = ggplot2::element_text(size = 10, angle = 0,
                                            colour = 'black'))
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
